requirements.txt
numpy==1.19.3
opencv-python
pytesseract
pillow
pywin32
pyautogui
mtranslate